pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pr_cost_estimator"
version = "0.1.0"
description = "Estimate the token and dollar cost of AI-based code review on a pull request."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "PR Cost Estimator Contributors" }
]
keywords = ["ai", "code-review", "cost", "tokens", "pull-request", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control",
    "Topic :: Utilities",
]
dependencies = [
    "tiktoken>=0.7.0",
    "gitpython>=3.1.40",
    "rich>=13.7.0",
    "requests>=2.31.0",
]

[project.scripts]
pr-cost-estimator = "pr_cost_estimator.cli:main"

[project.urls]
"Homepage" = "https://github.com/example/pr_cost_estimator"
"Bug Tracker" = "https://github.com/example/pr_cost_estimator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pr_cost_estimator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
